[sourcegen] Streamline dataclasses
