+++ interfaces/sourcegen/src/sourcegen/clib/generator.py
@@ -9,7 +9,7 @@ from dataclasses import dataclass
 
 from jinja2 import Environment, BaseLoader
 
-from .._dataclasses import HeaderFile, Param, ArgList, CFunc, Recipe
+from ..dataclasses import HeaderFile, Param, ArgList, CFunc, Recipe
 from ..generator import SourceGenerator
 from .._helpers import with_unpack_iter
