+++ interfaces/sourcegen/src/sourcegen/dataclasses.py
